requirements.V1.txt
pandas==2.2.0
numpy==1.26.0
scikit-learn==1.4.0
streamlit==1.32.0
plotly==5.18.0
python-dotenv==1.0.0
requests==2.31.0
geopy==2.4.1
xgboost==2.0.3
lightgbm==4.3.0
shap==0.44.1
seaborn==0.13.2
matplotlib==3.8.3
